+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["Cacheable", "flexbox", "persistor"]
+  "cSpell.words": ["Cacheable", "flexbox", "Formik", "persistor"]
 }
