+++ src/components/UI-kit/form/inputFile/InputFile.styled.ts
@@ -1,3 +1,4 @@
+import { Field } from 'formik';
 import styled from 'styled-components';
 
 export const LabelAddPhoto = styled.label`
@@ -13,11 +14,19 @@ export const LabelAddPhoto = styled.label`
   }
 `;
 
-export const InputAddPhoto = styled.input`
-  width: 0.1px;
-  height: 0.1px;
+export const InputAddPhoto = styled(Field)<{
+  border: string;
+  backgroundColor: string;
+}>`
+  width: 16px;
+  height: 16px;
   opacity: 0;
   overflow: hidden;
-  position: absolute;
-  z-index: -1;
+  /* position: absolute; */
+  /* z-index: -1; */
+`;
+
+export const ErrorText = styled.p`
+  color: red;
+  font-size: ${p => p.theme.fontSizes.xs};
 `;
